[Aptos Compression] Extract client into separate file.
